+++ src/test/java/com/avioconsulting/mule/testing/invocation/listeners/ListenersEnabledTest.groovy
@@ -39,7 +39,6 @@ class ListenersEnabledTest extends
     @Override
     List<String> getConfigResources() {
         [
-                'global-test.xml',
                 'listener-enabledisable-test.xml'
         ]
     }